requirements.txt
Flask==2.2.2
pydantic==1.10.2
mypy==0.982
wemake-python-styleguide==0.17.0
SQLAlchemy==1.4.41
psycopg2-binary==2.9.4
python-telegram-bot==13.14
httpx
